Fix for wrong workspace when running both -w -p at the same time
The program now runs the workspace selection before the project
selection, since it needs to know what workspace id to use for the
project.
The way that the JSON file gets populated was changed. The
PROJECTS area of the JSON file now looks like this:
```
"PROJECTS": {
	WORKSPACE_ID: {
		PROJECT_ID: PROJECT_NAME,
		...
	},
	WORKSPACE_ID: {...}
}
```

---

main.py
- The workspace id part of the code in the command_start function was
moved above the project id code.
- The project_selection function in utils.py now asks for workspace_id.

timers.py
- The start_timer function now has the workspace_id parameter before
project_id just so that it is consistent in all the functions.

utils.py
- Change add_projects_to_config function to configure the config.json
file in the way documented above.
- Add 'WORKSPACES' to the delete_user_data function so that it actually
removes that data when running `togglcli setup` to reconfigure.
- Add workspace_id parameter in project_selection function.

---

Resolves: #6