Add pause and resume timer functionality
The user can now use `togglcli pause` and `togglcli resume` to
"pause/resume" a timer. What the program is actually doing is stopping
the timer and saving all the data of that timer and then starting a new
timer with that same data.

---

main.py
- Add billable argument for `togglcli start`. Feature only available for
Toggl Pro accounts, so it defaults to False (not billable) for those
users.
- Add pause/resume argparse commands and functions.

timers.py
- Add billable parameter to start_timer() and data dictionary for the
post request.
- Add for_resume code in stop_timer() to save the data when pausing
a timer.
-Add resume_timer() function to, as it says, resume a paused timer.

utils.py
- Add add_previous_timer_to_config() function to run when pausing a
timer.
- Add remove_previous_timer_from_config() function to, as it says,
remove the previous timer data from the json file.

config.json
- Add empty "PREVIOUS_TIMER" dictionary.