Add no project option to `-p` argument.
Add code to select no project when starting a new timer with the `-p`
argument. Previously, the user could only select a project, but if the
user realizes that they don't acually need a project then they would be
forced to use `Ctrl + C` to cancel the command. Now there is an option
`0: Don't use any project` that the user can use.

***

Changes:
utils.py -
    project_selection() - Print out `0: Don't use any project` and check
if user enter that selection. This makes the `project_id = ""`

timers.py -
    stop_timer() - Noticed a typo and fixed it.

***

Resolves: #9