Add create project option.
Add code to let the user create a new project from the CLI. The user can
now use `tgl create project "The name of the project"` to create a
project. If the user has multiple workspaces, the program will ask what
workspace the user want's to create the project on. If the user only has
one workspace, the the program will use that one without asking the user
anything about it.

***

Changes:
main.py -
    Add `create` parser that takes in a `request` (a choice for what to
create, in this case project) and the name to assign to that request.
    Add `command_create()` function to process the parser's arguments.

timers.py -
    Add `create_project()` function to post the request to Toggl's API.

utils.py -
    Modify `workspace_selection()` function to not print out a verbose
method when the `verbose` argument is False.
    Modify some spacing issues in the outputs of the same function

***

Resolves: #10